#35 ♻️ refactor: 컴포넌트 export 방식 수정
